+++ art_web_app/utils.py
@@ -1,14 +1,6 @@
 from artworks.models import Artwork, Category
 from users.models import AuthUserModel
-
-
-def get_next_position(model, **kwargs):
-    try:
-        next_position = model.objects.filter(**kwargs).order_by('-position').first().position + 1
-    except AttributeError:
-        next_position = 1
-    
-    return next_position
+from galleries.models import Gallery
 
 
 def _split_queries(user_query):
@@ -31,6 +23,7 @@ def _split_queries(user_query):
 def get_query(query=None):
     art_query_results = set()
     user_query_results = set()
+    gallery_query_results = set()
     queries = _split_queries(query)
     if len(queries) > 0:
         for q in queries:
@@ -41,8 +34,11 @@ def get_query(query=None):
                 continue
             artworks = Artwork.objects.filter(title__icontains=q)
             users = AuthUserModel.objects.filter(username__icontains=q)
+            galleries = Gallery.objects.filter(name__icontains=q)
             for art in artworks:
                 art_query_results.add(art)
             for user in users:
                 user_query_results.add(user)
-    return art_query_results, user_query_results
+            for gallery in galleries:
+                gallery_query_results.add(gallery)
+    return art_query_results, user_query_results, gallery_query_results
