+++ artworks/templates/artworks/art-card.html
@@ -1,1 +1,6 @@
-<img class='img-result' src="{{ art.thumbnail.url }}" alt='{{ art.title }}'/>
+<div class="swipe-card">
+    <a class='img-link' href="{% url 'artworks:view' art.id %}"><img src="{{ art.image.url }}" alt="{{ art.title }}"></a>
+    <div>
+        <p class='likes-counter'>{{ art.likes.count }} Likes</p>
+    </div>   
+</div>
