+++ templates/gallery-card.html
@@ -1,0 +1,15 @@
+<div class='profile-galleries'>
+        <a class="gallery-link" href="{% url 'galleries:view' gallery.id %}">
+            <div class="gallery-card-container"> 
+                <div class='gallery-card'>
+                    {% for art in gallery.artworks.all|slice:":3" %}
+                        <img class="gallery-art" src="{{ art.image.url }}" alt="{{ art }}">
+                    {% endfor %}
+                    {% if gallery.artworks.all|length > 3 %}
+                        <span class="gallery-plus">+ {{ gallery.artworks.count|add:"-3"}}</span>
+                    {% endif %}
+                </div>
+                <p>{{ gallery }}</p>
+            </div>
+        </a>
+</div>
