--- utils/__init__.py
@@ -1,1 +1,0 @@
-from .get_utils import get_query
